requirements.txt
ccxt==4.1.92
python-dotenv==1.0.0
requests==2.31.0
colorama==0.4.6

